requirements.txt
fastapi
uvicorn
requests
python-dotenv
pillow
transformers
torch
beautifulsoup4
selenium 
webdriver-manager
transformers 
opencv-python
pytrends
gdown
playwright
nest_asyncio
beautifulsoup4